Dynamic syntax highlighting  (#35)
* Dynamically-updating sample XAML works again

Get rid of the misleading <RichTextBlock> which are never
shown to the user. Those are left over from earlier versions
of the sample. Instead, there is a Xaml property whose value
is a string that is the sample XAML. Dynamic content in the
string is represented by $(...). Accompanying the string
is a collection of ControlExampleSubstitution objects, each
of which has a Key and a Value. The key is the thing that
goes between the parentheses, and the Value is the replacement.
The Value can be any object, and the code will try to
pretty-print it in a XAML-friendly way:

* SolidColorBrush is represented by the underlying color.
* Null is represented by the empty string.
* Everything else becomes its own ToString().

Added a CSharp string property for sample C# code.
Substitutions are performed in the Xaml, CSharp, and
in the corresponding source files (XamlSource, CSharpSource).

Other cleanup:

* Removed unused converters.
* Fixed hyperlink in ConnectedAnimation.
* Image page erroneously showed markup for Example3 as part
  of Example2.
* RevealPage sample markup didn't match actual markup.
* SoundPage erroneously rendered its C# code as XAML.
* TreeView markup was incorrect.
* Fix TreeView typo "ExpolrerItemTemplateSelector".
* ViewBox.StretchDirection, ScrollViewer.ZoomMode, and
  CalendarView.Language, and implicit transitions Opacity
  and Scale now update correctly.
* Many indentation errors.

Pre-existing issues not fixed:
* ConnectedAnimations: Changing the Configuration does not
  update the sample code.
* EasingFunctions: EasingMode is not included in the markup.
* ImplicitTransitions AnimationFlags are not hooked up.
* InputValidationPage and RepeaterPage are orphaned pages.

* Get rid of unused converters, simplify substitution collections, pick up some orphaned pages (CompactSpacingPage, RepeaterPage)

* Fix line terminators and whitespace

* Fixing the PersonPicturePage to have dynamic updating. We needed to add Mode=OneWay.