Update ListViewSample2_xaml.txt
I spent a view hours trying to figure out groups, and it wasn't until I looked at the source code for the examples that I saw the "GroupStyle" bit that was required. You have to have it, and unless you know is there from the source, the example doesn't give enough to replicate it.