This is port of internal fix which adds Confirmation of Action feature to many actions in WinUI gallery (#1245)
Confirmation of Action is an accessibility feature where visual changes are notified to the impaired user using audio notification like Narrator. The common actions where it can be used when a visual is changing color or app theme is changing from light to dark or vice versa.

This PR introduces a AnnounceActionForAccessibility helper function which helps create this contextual announcement. This announcement is used by Narrator to inform the user that a large visual change has happened like scale/rotate of an object.
---------

Co-authored-by: Pratik Anand <praanan@microsoft.com>