Added a "Capture element / camera preview" page in the Media section. (#1357)
## Description
`CaptureElement` from UWP XAML does not exist in WinAppSDK, but
`MediaPlayerElement` can be used in its place. The change adds a new
sample page to demonstrate how this can be done.

Note: Currently this page just reuses the same icon as the
MediaPlayerElement page. Ideally at some point we'll get a camera icon
which could be used here instead.

## Motivation and Context
Demonstrate the couple of calls to get from a `MediaCapture` to an
`IMediaPlaybackSource` necessary to display the capture in a
MediaPlayerElement. This also demonstrates basic creation of a
`MediaCapture` to provide an end-to-end sample. This sample should help
address questions in microsoft/microsoft-ui-xaml#4710 and
microsoft/microsoft-ui-xaml#8214.

## How Has This Been Tested?
Ad-hoc testing and testing with Accessibility Insights for Windows tool.

## Types of changes
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)