Remove WindowHelper from FilePicker sample code (#1396)
## Description
Removed the WinUI3Helper class from the `FilePicker` sample code and
showed how to get the `MainWindow` from the `App` class instead.

## Motivation and Context
The sample code for the `FilePicker` controls uses the WindowHelper
class which is implemented as a helper class in this project. This is
very confusing because it looks like the WindowHelper class belongs to
WinUI 3 and when one tries to use that code in their apps it just
doesn't work. I find several questions in StackOverflow related to this
and I think it's better not to use the WindowHelper in sample code.