Searching icons by tags (#1562)
<!--- Provide a general summary of your changes in the Title above -->
## Description
<!--- Describe your changes in detail -->
Closes #1544.
This PR will enable searching icons based on tags on the Iconography
page. Icon tags are assigned on the IconsData.json file.

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->
The current icon searching is based only on the icon name or its code.
Some icon's names are not very intuitive, or some icons can represent
several things. So, to get an icon code and if you don't know the icon
name, you need to search it manually by scrolling through more than 1400
icons.

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->
Tested on the VS.

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)