[SceneGraph LKG] Update capture to use lifted VisualCapture instead of RenderTargetBitmap (#1655)
Using some Composition native interop, we can read the bits back out of
a CompositionDrawingSurface. Using the lifted VisualCapture API, we are
able to get crisp text out of our captures, even at high DPI.

This requires a little bit of smoke and mirrors where we scale the UI up
before the capture. We cover up this one-frame scale with a temporary
VisualSurface.

Add a C++/WinRT component project and C# projection project to handle
the Composition native interop.

Some small tweaks to the twirl dismiss.

---------

Co-authored-by: Ian Blaauw <iablaauw@microsoft.com>