Fix: `Color > HighContrast` theme toggle highlight (#1680)
<!--- Provide a general summary of your changes in the Title above -->

## Description
The changes in this PR prevent the `SampleThemeListener` from applying
theme changes to all child elements, including some textboxes, which do
not need to be affected by the theme toggle. These changes address the
issue of inconsistent text highlighting in the "High Contrast" section
of the `ColorPage` by improving theme handling. The main changes are:
- Removed `<SampleThemeListener>` from `ColorPage`.
- Added `<SampleThemeListener>` to other "Color Sections" pages as
needed for better modularity.



## Motivation and Context
Fixes #1671 

## How Has This Been Tested?
**Manually Tested** by toggling the theme in all sections of the
`ColorPage` and observing the behavior.

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)