Enhance search functionality by adding tags to icons (#1706)
<!--- Provide a general summary of your changes in the Title above -->

## Description
This update adds tags to the icons listed in `IconsData.json`, improving
the searchability of icons on the iconography page. The only change is
the addition of tags to the `IconsData.json` file.

**How the Tags Were Chosen:**  
- For many icons, the tags are derived from those used in the [Fluent UI
System Icons](https://github.com/microsoft/fluentui-system-icons.git)
project, particularly for common icons.
- For most icons and icons that not included in the Fluent UI System
Icons, tags were added through a combination of personal effort and
assistance from AI.

## Motivation and Context
- Many icons currently lack tags, limiting the effectiveness of this
feature. Adding tags to all icons ensures that the search functionality
is fully utilized, making it easier for users to find the icons they
need efficiently. This enhancement aligns with the goal of providing a
seamless and user-friendly experience.
- Closes #1705 
- Related to #1544, #1562 and #1693

## How Has This Been Tested?
**Manually tested**

## Screenshots

![image](https://github.com/user-attachments/assets/6d289244-6996-4a4f-8c7b-5c51f58fb5cc)

![image](https://github.com/user-attachments/assets/9f9b6d9e-0829-474e-b712-07296903076a)


## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)