Remove the unicode point & the duplicate icon name from icon details (#1724)
## Description
This PR includes the following updates to improve the clarity and
conciseness of the icon details:
1. **Remove Unicode Point from Icon Details**: The `unicode point` has
been removed, as it adds redundancy without offering additional value.
The primary ways to use font icons (`text glyph` in XAML and `code
glyph` in C#) are already displayed.
2. **Remove Duplicate Icon Name**: Eliminates the repeated display of
the icon name within the icon details, streamlining the information
presented.

## Motivation and Context
- This change is required to streamline the icon details for clarity and
conciseness, especially with #1693 on the way.
- Closes #1722 

## How Has This Been Tested?
**Manually tested**

## Screenshots (if appropriate):

![image](https://github.com/user-attachments/assets/69f6fe64-aaa1-4717-b83b-1e02c5638260)

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)