ScratchPadPage.xaml was broken, Grid was blocked (#1759)
## Description

I noticed that the interactive area of the page did not receive any
mouse input. Keyboard did work. I saw that conversion for
`RowDefinitions="*, Auto"` was applied to the incorrect (child) Grid,
causing the TextBlock to be in the same Grid cell as the RichInput.

## Motivation and Context

Fix blocked input

## How Has This Been Tested?

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

---------

Co-authored-by: Niels Laute <niels.laute@live.nl>