Update the spacing in the sample code display of the `ControlExample` control (#1789)
## Description
This PR eliminates the unnecessary top and bottom space occupied by the
sample code in the source code expander. The only change is updating the
`Padding` of the `Expander` element to improve layout and content
spacing.

## Motivation and Context
For a better and cleaner UX.

## How Has This Been Tested?
Manually tested

## Screenshots (if appropriate):
**Before**

![image](https://github.com/user-attachments/assets/e6113363-7235-407f-8ff8-ec828328579e)

**After**

![image](https://github.com/user-attachments/assets/70f67a1e-3778-499a-9a91-78ecf902b5ab)

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)