Improve `SampleCodePresenter` layout and `CopyCodeButton` behavior (#1799)
## Description
This update improves the `SampleCodePresenter` by changing the position
of the `CopyCodeButton`. Now, it is in the same column as the
`CodePresenter`,making it appear as a floating element instead of
occupying unnecessary space. The padding and margins on the left and
right sides have also been adjusted. The button’s visibility now changes
when the user selects text or moves the horizontal scrollbar.

## Motivation and Context
Before this change, the `CopyCodeButton` was in a different column,
which wasted space. Moving it into the same column as the
`CodePresenter` makes the layout better and more useful. The button now
appears only when needed, making the experience smoother. This design
was inspired by GitHub’s code presenter.

- Related to
https://github.com/microsoft/WinUI-Gallery/pull/1789#issuecomment-2708231442

## How Has This Been Tested?
- Manually tested different code samples.

## Screenshots (if appropriate):
### Spacing before and after changes
#### Before

![image](https://github.com/user-attachments/assets/03343f8e-9d11-4844-88fc-acfd0a879e4d)
#### After

![image](https://github.com/user-attachments/assets/b089df4e-4fd7-4aa2-b5b6-8178d58a2f59)

![image](https://github.com/user-attachments/assets/4a8c9ee3-8ade-436e-b466-6448ae4b60a2)
### GitHub code presenter behaviour and the implemented behaviour
#### GitHub behaviour

![GH1](https://github.com/user-attachments/assets/bcaea287-b605-43fe-8a6b-3fc876774eb1)
![GH2](https://github.com/user-attachments/assets/9406d50c-a8da-47d8-b841-a8bd9ad3d897)
#### The implemented behaviour

![1](https://github.com/user-attachments/assets/30658288-1baa-4c01-9bb9-8b846403b08a)
![2](https://github.com/user-attachments/assets/f0a18480-e989-4eb0-9981-ad361dc13d7f)

## Types of changes
- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

---------

Co-authored-by: Niels Laute <niels.laute@live.nl>