Fixed the drag and drop between ListViews (#1898)
## Description
The current drag and drop between ListViews continually duplicates the
contact being dragged and does not remove the contact from the original
list. This fix ensures that a contact will only show up in one of the
ListViews at any given time.

## Motivation and Context
Developers should have an example of a working drag and drop between two
ListViews

## How Has This Been Tested?
This logic has been manually tested to verify that a contact can be
moved between ListViews.

## Screenshots (if appropriate):

![image](https://github.com/user-attachments/assets/bfaa8593-c8d2-42e3-a007-2ac274511d78)

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

---------

Co-authored-by: Niels Laute <niels.laute@live.nl>
Co-authored-by: Marcel Wagner <marcelwagner@microsoft.com>
Co-authored-by: Marcel W. <marcel.alex.wagner@outlook.com>
Co-authored-by: Zakaria Tahri <zakotahri@outlook.com>