Fix: sync titlebar button colors with app theme changes (#2016)
## Description
When the app theme is set to Light or Dark (not “Use system setting”)
and the app theme differs from the system theme, the app initially opens
in the system theme and then switches to the chosen app theme.
Previously, the titlebar button colors were not updated during this
theme change, resulting in low contrast.
This fix updates the button colors whenever the root grid theme changes
and removes the old broken logic.

## Motivation and Context
- Fixes #1605 
- Fixes #1970 

## How Has This Been Tested?
Manually tested

## Screenshots (if appropriate):

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)