Fix light theme caption button hover, cleanup TitleBarHelper (#2026)
This PR fixes an issue where setting the app to light mode while system
theme is dark would result in over for caption buttons being white on
white.

As part of this, I cleaned up the TitleBar helper and switched to one
function to be called by everyone instead of the segmented functions.