Child window exception fix (#2007)
<!--- Provide a general summary of your changes in the Title above -->

## Description
<!--- Describe your changes in detail -->
This PR replaces `Win32WindowHelper` with `OverlappedPresenter` to set
window max size.

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
The `Win32WindowHelper` relied on static fields, which led to app
crashes when handling multiple windows. The `OverlappedPresenter`, as
demonstrated in the AppWindow sample page, is the recommended approach
to setting window size constraints.

<!--- If it fixes an open issue, please link to the issue here. -->
Note: This PR supersedes #2005, which was closed due to an incomplete
approach.
Fixes: #2003

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->
Ran the app and verified the changes.

## Screenshots (if appropriate):

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to change)

---------

Co-authored-by: Niels Laute <niels.laute@live.nl>